1. refactor: reservation page dir update 2. reservation page complete
