+++ src/router/RootRouter.tsx
@@ -1,6 +1,7 @@
 import App from '@/App';
 import RootLayout from '@/components/Layout/RootLayout/RootLayout';
-import DetailReservation from '@/pages/Rservation/DetatilReservation/DetailReservation';
+import DetailReservation from '@/pages/Reservation/Detail/DetailReservation';
+import ReservationIndexPage from '@/pages/Reservation/Index/Index';
 import {
   createBrowserRouter,
   RouterProvider,
@@ -12,7 +13,7 @@ export const RootRouter = () => {
   const router = createBrowserRouter(
     createRoutesFromElements(
       <Route path="/" element={<RootLayout />}>
-        <Route path="reservation" element={<App />} />
+        <Route path="reservation" element={<ReservationIndexPage />} />
         <Route path="reservation/*" element={<DetailReservation />} />
       </Route>
     )
