+++ src/styles/global.scss
@@ -1,4 +1,5 @@
 @import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/variable/pretendardvariable-dynamic-subset.min.css');
+@import '@/styles/theme.scss';
 
 html,
 body,
@@ -41,4 +42,8 @@ input {
 button {
   border: none;
   cursor: pointer;
+}
+
+.Divider {
+  border: 1px solid $gray300;
 }
